Now avoiding use of the bit firmware module
The bit math can be done without the module, which saves a few bytes of ram for each module affected.

It also removes the requirement for the bit module to be in the firmware.

Code is harder to read as a result, so I’ve commented out the original code using the bit library and left it there.