Prevents CSV and logging files from getting too large
There's a bug in the firmware's SPIFFS filesystem, that leads to file
writes slowing down and eventually freezing as the filesystem gets
full, and leading to an unstable filesystem and potentially
unrecoverable data.

The problems start when the filesystem is ~75-80% full. Until the bug is
fixed in the firmware, writing to CSV or log files is prevented when
less than 25% of the filesystem space remains. For a 4MB module, this
means there is about 800 KB of unusable space, and only about 2.5 MB
that
is useable.